todo: refactor those utils as Model methods
